This commit implements the JWT Authentication Method
In this we have leveraged the libraries provided by the Springboot Framework, and to implement this we have also included certain dependencies like jsonwebtoken dependency in the Maven Configuration File.

Along with this we have implemented the two SecurityFilter bean, one which will just look for the Admin request, other one will look for all other endpoints request.